basic commutation mutation pass

+++ src/run.cpp
@@ -166,13 +166,13 @@ void Run::loop(){
                     current_generator->run_genetic(output_dir, n_programs.value_or(0));
 
                 } else if (run_mutate){
-                    for(int build_counter = 0; build_counter < n_programs.value_or(0); build_counter++){
+                    for(size_t build_counter = 0; build_counter < n_programs.value_or(0); build_counter++){
                         fs::path current_circuit_dir = output_dir / ("circuit" + std::to_string(build_counter));
                         current_generator->ast_to_equivalent_programs(current_circuit_dir);
                     }
             
                 } else {
-                    for(int build_counter = 0; build_counter < n_programs.value_or(0); build_counter++){
+                    for(size_t build_counter = 0; build_counter < n_programs.value_or(0); build_counter++){
                         fs::path current_circuit_dir = output_dir / ("circuit" + std::to_string(build_counter));
                         current_generator->ast_to_program(current_circuit_dir, std::nullopt);
                     }
